Make using old PIP easier
This change improves the diagnostics from setup.py. A new helpful
error messages will trigger if any dependencies are missing when
installing the package. This typically happens when PIP version is <19
and the user is missing the build dependencies.

Example error message:

    Missing module 'skbuild'. Please install 'scikit-build' manually
    or use PIP>=19 to handle build dependencies automatically (PEP
    517).

Version number is bumped to 0.9.1 and this version will be uploaded to
PyPI.

The README.md is now simplified to not use venv since we gracefully
handle older PIP version. It is also updated to install the package
from PyPI and not git.

Added more info to Versioning section of README.md

Added PyPI link to README.md

wip