Remove the documentation from the C++ wrapper layer.
This layer is not visible to the end user and documenting it is just
extra work.

The important documentaton is in the Python wrapper layer, in Python
files.

Soon we will add wrappers for the 2D API and the Eye Hand API. Then we
should not be required to document those additions twice, but rather
just add documentatoin to the Python files.