Releaseable point cloud
Instead of transforming the point cloud directly to a numpy buffer, we
should instead return a intermediate layer, since in the underlying C++
API point cloud exposes various functions such as width and height.

Making buffer releasable also created some technical debt:
assertNotReleased is needed since accessing the underlying a C++ object
after it has been released causes a crash.

Ignored flake8's D402, since it mistakenly thought that the paranthesis
in the docstring was the signature of the function.