Improve error message when module build does not work
This improve error messages during `pip install zivid` on systems
without binary packages and without C++17 support. This is implemented
as a nested exceptions where the outer most excetion gives a fix-it
hint.

I also improved formatting of the fix-it hint when modules are missing
and sorted some dependency lists.