Remove most of the hardcoded versions from pip requirements files
If we do not freeze the version we will automatically get the latest and
greatest of linting tools, formatting tools, etc. We would need to
update this repo whenever any linter tools get a new test we cannot
pass, but I think this is a worthwhile trade.

Build requirements should be frozen, so that building the code should
work. Test and lint requirements does not need to be frozen.

The newer pip modules required some changes:
    - pytest-unmarked was replaced with calling pytest with -m
    - added Raises section in docstrings
    - ignoring pytest import-outside-toplevel
    - change error type to TypeError when initializing camera, etc