Restructure requirements files
Prework for updating the pip packages. The latest version of pylint
requires that it knows about various plugins for pytest in order to
successfully lint. This commit makes it easier to install the packages
for a specific task, so it is now easier to add the test packages before
linting.