Restructure pipeline workflow
Prework for updating the pip packages. The latest version of pylint
requires that it knows about various plugins for pytest in order to
successfully lint. This demand requires the dependencies of setup, test
and lint to be separated, since before this commit all of the steps were
more or less done in one big blob.

By restructuring the azure-pipeline worklfow into setup, build and
test/lint it is much easier to reason about the pipeline, and now easier
to install the correct dependency for the correct stage.