Enable Windows CI tests
In order to run tests on Windows, Intel OpenCL runtime needs to be
installed to run the tests on a compute device. It is also required to
create a configuration file so that the tests are run on the CPU instead
of a GPU (since there are no GPU available in azure-pipelines)

Added update environment script in order to load the latest environment
variables set by `Zivid-SDK` after successful installation. `PATH`
environment variable must be updated befre running the tests.

Ignored sample testing for windows, because of the following error:
```
python: can't open file 'D:\a\samples\sample_print_version_info.py': [Errno 2] No such file or directory
```

Ignored `winreg` import error, this module is not available on Linux