Simplify tests to only use file-camera
This has the following benefits:
- Simplifies the testing code (conftest.py / sample_data.py). Most tests
now only depend on a single file, FileCameraZividOne.zfg, now commited
to this repo with LFS. We are no longer dependent on an external
downloaded sample zip-file.

- Due to the test-Frame now coming from a file camera, as opposed to an
old ZDF file, tests on frames and point clouds now by default test the
results of the current processing pipeline. This makes it possible to
actually test newer features such as frame.settings without getting an
exception because the Frame is from an outdated SDK version.