CI: Fix Python version selection for Windows pipelines
For a long time there has been a bug in the Python version selection for
Windows pipelines in azure-pipelines.yml.  This commit fixes this by
adding a "UsePythonVersion", as described in
https://docs.microsoft.com/en-us/azure/devops/pipelines/ecosystems/python

This commit also turns on the --verbose flag for pip in CI builds, in
order to get the kind of output that would have made this bug easier to
catch.

By doing this it was discovered that we cannot actually support Python
3.5 for Windows without further adjustments to the build system.  Since
Python 3.5 is end-of-life, we choose to remove support for 3.5 on
Windows instead.