Fix to work with Python 3.8/3.9 on Windows
Before this commit, "import _zivid" would fail with
```
ImportError: DLL load failed while importing _zivid: The specified module could not be found.
```
on Windows with Python 3.8 and 3.9 due to a new .dll search mechanism
that was introduced. This commit adds a workaround for this, which
allows us to support these Python versions on Windows. The workaround
also has an accompanying comment that gives further details.