Small adjustment in README
It is a bit misleading to state that zivid-python supports "2.2.0 or
newer". If SDK 2.3.0 is installed then `pip install zivid` will fail,
either immediately (Ubuntu) or at runtime (Windows). While it is true
that one could check out the repository, edit `setup.py` to target
2.3.0, and then build from source, this is an uncommon use-case.

Since the current text was misleading for the majority of users that
merely does `pip install zivid`, we here remove the "or newer" part.
This should make it more likely that the common user installs the
correct SDK version first.