Migrate CI from Azure Pipelines to Github Actions
This commit migrates the build+test+lint CI workflow from Azure
Pipelines to Github Actions. This is done in preparation for expanding
to more advanced workflows for e.g. deployment, since we know Github
Actions better as of now.

This commit also adds unit-testing on Archlinux, so that we may get
early warning if we ever stop working with the latest Python.