Add Github Actions workflow to assist release deployment
Previously the CI workflow did build and test in the same step, and did
not create any deployable artifacts.  This commit changes the CI
workflow to create deployable artifacts for PyPI (or direct installation
with pip). It also separates the creation and testing of these artifacts
into separate jobs.  The workflow does not go as far as actually upload
to PyPI. Instead it provides, as an artifact, a zip-file which contains
the full set of tested and validated wheels/source-dists ready for
upload to PyPI. A developer with the necessary credentials may then
simply download the artifact called "dist_all" and run `twine upload *`
on the unzipped contents.