Fix bug where __bool__ -> True when valid() -> False
The classes in the calibration API have the following implementation on
the front-end,
```
def __bool__(self):
    return bool(self.__impl)
```
which was intended to represent the corresponding boolean operator on
the C++ API level. However, the boolean operator of the `__impl`s were
never overridden on the Pybind11 level, so they instead defaulted to
some behavior that seemingly always returns True, and is independent of
the `operator bool` in the C++ API. This caused the following kind of
bugs in the Python API
```
detection_result.valid() # False
bool(detection_result)   # True
```
This commit fixes this by explicitly hooking up the `__bool__` of the
Pybind11-generated classes to the `operator bool` of the C++ classes.
This yields behavior that is identical to the C++ API, as intended.

No unit-tests are added for this because instantiating invalid
calibration objects like DetectionResult requires running
detect_feature_points with a non-checkerboard point-cloud, and this
would unreasonably slow down the CI.