Use `pip-compile` to manage python dependencies
Makes it easier to bump dependencies in the future, as one can simply
run `pip-compile --upgrade /path/to/requirement_file.in` to bump them.

This commit also bumps conan, which is required to make the CI green
again. The reason for this is to be
`Let's encrypt root authority server shutdown on September 30th`
See https://github.com/conan-io/conan/issues/9695 for more information.
The latest conan version has the correct certificates.