Bump SDK version from 2.4.2 to 2.5.0
This automatically extends the CameraInfo class to include the new
`CameraInfo.Model` enum type. This is the first time a data-model has
six members at the same level of its hierarchy, so it triggers pylint on
`too-many-branches` and `too-many-boolean-expressions`. This is not a
problem for the generated data-model front-ends, so we add exceptions
for these two in the `datamodel_frontend_generator.py`.