Revert "Added python 3.10 testing for Linux using the Fedora 35 image"
The unit-test "test_capture_from_file" on fedora:35 (python3.10)
suddenly started consistently failing. After some investigation it was
found that it comes from the following error at the time when the file
camera is destroyed:

```
pure virtual method called
terminate called without an active exception
Aborted (core dumped)
```

In order to get the CI green again we remove Fedora 35 from the support
matrix for now. The issue #161 was created for fixing this issue and
re-adding it.

This reverts commit 8c703a8e914e564a0204b3d0878e9a3d5e7aea18.