Bump pybind11 from 2.6.1 to 2.9.1
This allows nested exceptions from C++ to be properly handled, so that
the user can see more than the top-level exception.

The version bump required a small adjustment to the datamodel frontend
generator in order to yield the same frontend python code as before, so
that is also done in this commit.