Wrapper: Cleanup runs in the python world in addition to libc world
A Zivid::Application may depend on and share internal static variables
from third party dependencies. Using a Singleton<Application> creates a
race in the order of destruction of statics. This order is deterministic
if Zivid::Application's constructor also depended on these internal
static variables but that is not always the case. This commit releases
singletons on module unload.