Disable unit-testing on Archlinux
The unit-tests fail with SDK 2.6 in combination with the latest
Archlinux. This commit temporarily disables this in order to get the CI
green again. There is evidence that this will be improved with SDK 2.7,
and at that point we should re-enable Archlinux and Fedora35 testing.
This is tracked by issue #161.

(cherry picked from commit b948a784c9f77d26cd95491e11c0d4c92aa48da0)