Build and test for Windows+Python3.11
This commit adds Windows+Python3.11 to the CI workflow, which means that
a `.whl` file for that system is built and tested.

Supporting Python 3.11 required the following changes:
- Bump `pybind11`  to version 2.10.0
- Bump `pytest` to version 7.2
- Remove `pytest-pythonpath` since it is deprecated and no longer needed
- Bump `black` to version 22 to keep compatible with new `pytest`.

Since Python 3.6 has been end-of-life since December 2021, we remove the
Windows+Python3.6 combination at the same time. An exception is made for
Ubuntu 18.04 (which has Python 3.6 as its default Python), since Ubuntu
18.04 is not yet end-of-life. Unfortunately the new `pytest` is not
compatible with Python 3.6, so we had to make a separate "legacy"
version of `test.txt` requirements file for this CI job.