Pin conan version
Conan 2.0 was released to PyPI today and breaks `zivid-python` build
with the following error message due to a change in Conan's API:

```
conan install: error: unrecognized arguments: --install-folder
```

This change pins Conan to 1.x and fixes the issue while we figure out
how to bump to Conan 2.x.

In order to allow a new `zivid-python` release to PyPI we also bump its
patch-version.