Update Arch AUR package references
Recently, new versions of the Arch AUR packages for SDK 2.9 were
uploaded. These packages reference the new `downloads.zivid.com` domain
instead of the temporary one.

This commit updates the CI in the zivid-python repo to use those new AUR
packages, so that we do not need to maintain the temporary URLs.