Make CI independent of Arch Linux
It has been decided that we no longer want the CI workflow to depend on
Arch Linux, in order to avoid the workflow suddenly breaking due to
upstream changes.

This commit does the following:
- Migrates the linting job to use Ubuntu instead of Arch.
- Removes the Arch based build & test job.

This also has the positive effect that bumping the supported SDK version
no longer depends on the intermediate step of releasing to Arch AUR.

From now on the linting job should be run on the newest Ubuntu version
possible. Supporting 23.04 will require some more work, so for now we
run it on 22.10.