Redefine the Linux build-and-test matrix in the CI
It has been decided that we should cover the following in the Linux
build-and-test matrix:

- All Ubuntu versions officially supported by Zivid SDK
- All Ubuntu versions newer than the latest LTS
- The three latest Fedora releases that we can support

For now, due to issues with the OpenCL CPU runtime we use, the newest
Ubuntu and Fedora that we can support in this CI are 22.10 and 36,
respectively, not the latest releases. This issue can be dealt with as a
follow-up task.