Add test job for Fedora 37 (Python 3.11)
Since we started to use a more modern OpenCL CPU runtime we are now able
to support and test on Fedora 37. We have decided to only test on the
three latest Fedora releases, so therefore we remove the Fedora 34 CI
job.

Note: Fedora 38 is out since April 2023, but it appears that the Intel
OpenCL runtime packages do not support it yet.