Make Linux CI scripts use Python venv
As of version 23.04, Ubuntu no longer allows the user to use pip to
install packages into the system Python. When trying to do so, an error
message is shown telling the uses to use a venv instead.

In preparation for Ubuntu 23.04 support, and since it's good practice
anyway, this commit change the CI scripts to use Python venv instead of
system Python.