Fix error message in constructors
In many of the classes that wrap the C++ bindings, the constructor
checks the type of the argument and if the type is not correct, the
constructor raises an error:

"Unsupported argument for type impl: Got ... , expected ..."

Previously, the expected type was obtained by calling type() with the
binding class. However, the use of type() here will not give the name of
the binding class as wanted, instead it results in something like:

"<class 'pybind11_builtins.pybind11_type'>"

To fix this, this commit removes the unnecessary use of type().

MISC