Wrappers: fix toSnakeCase for IPV4
Previously, the `toSnakeCase` function used in the Wrappers would
convert `IPV4` to `ip_v4`. This conflicts with the `_to_snake_case`
function used in the datamodel code generator, which would return
`ipv4`. The latter is the version used in the Zivid SDK, so this commit
fixes the `toSnakeCase` function so that it uses the latter version.

More specifically, the `toSnakeCase` function would previously insert a
"_" in the following two cases:
  1. If an upper case letter preceded by a lower case letter is found.
  2. If an upper case letter is found that is preceded by an upper case
     letter and succeeded by a character that is not an upper case
     letter.

This commit adjusts the second case to also check that the succeeding
character is not a digit.