Unit testing: avoid mutating shared fixtures
We are using module scoped fixtures to speed up our testing, as some of
the fixtures (for example, Application and Camera) are expensive to
create. The downside of this is that some of our tests want to put these
objects into different states that conflict with other tests
(disconnecting a Camera, for example).

To prevent this issue, this commit makes the "file_camera" fixture
function scoped and adds a "shared_file_camera" fixture that can be
used by tests that don't need to mutate the fixture. This commit also
moves tests of the initialization and teardown of the Application class
into a separate module so that it doesn't conflict with other tests that
use the "application" fixture.