Unit testing: randomize test execution order
A recent commit 5b3f385 fixed an issue that was found in existing tests
as a result of adding a new test. This issue could have been caught
earlier by randomizing the execution order of the unit tests.

This commit adds pytest-randomly to the test requirements so that test
execution will be randomized.