Revert: "drop support for Python 3.7"
Support for Python 3.7 was removed in 47fdcd5, but we have decided to
revert this change and continue supporting it for now, as a workaround
was found for the issue that motivated dropping support.

Python 3.7 support is likely to be removed again in the near future
since it has reached EOL.

In the future we will update our policy on Python version support to
clarify when support for an old version ends. The current policy can be
found here:
https://support.zivid.com/en/latest/api-reference/sdk-lifecycle.html