Support for Zivid SDK 2.11.0
Adds support for Zivid SDK 2.11.0 for python wrappers.

This adds following camera states:
- Network.IPV4 that represents IPv4 address for network cameras

- Status enumeration to represent different camera states. Zivid SDK
  2.11.0 lists all the discovered zivid cameras as opposed to previous
  SDK behavior where only available zivid cameras were listed. Hence
  following new status are included:
`inaccessible`: The camera was discovered, but the SDK is not able to connect to this camera.
`busy`: The camera is currently in use by another process.
`firmwareUpdateRequired`: The camera is accessible, but requires a firmware update before you can connect to it.
`updatingFirmware`: The camera firmware is currently being updated in the current process.
`available`: The camera is available for use by the current process.
`connecting`: The camera is currently connecting in the current process.
`connected`: The camera is connected in the current process.
`disconnecting`: The camera is currently disconnecting in the current process.
`disappeared`: The camera was found earlier, but it can no longer be found.

- New subsampling setting: `Settings.Sampling.Pixel.blueSubsample4x4`,
  `Settings.Sampling.Pixel.redSubsample4x4`.