Upgrade to Conan 2
Conan 1 is being deprecated, and so it is necessary to upgrade to
Conan 2. In Conan 2, the CMake integration has changed. It is now
designed to function as a custom dependency provider for find_package
that reads a conanfile to determine which dependencies are required.
This requires us to bump minimum CMake version to 3.24.

This commit also bumps the conan dependencies to the latest version:
 - Eigen3: 3.4.0
 - Pybind11: 2.12.0