Add new non-experimental calibration board detection functions
This commit adds wrappers for newe calibration board detection functions
in the non-experimental `calibration` module. The following functions
are added:

`detect_calibration_board` - This function can be used with either a
Frame or Camera as input, and returns a DetectionResult.

`capture_calibration_board` - This function can be used to capture a
frame with a Camera that later can be passed as input to the former
function.