Bump to SDK 2.14.0
This commit bumps to the latest version of the Zivid SDK, 2.14.0 and
regenerates the datamodels. New features will be implemented in
following commits.

This commit adds a TypeName specialization for nested datamodel.
TypeName returns the name of the Python type that represents a data
model node. For a nested datamodel this is the wrapped class of
the nested datamodel.

This commit also updates checkerboard ZDFs and multicamera transforms.
In 2.14, we dropped support for legacy checkerboards that lack Aruco
markers. This commit replaces the checkerboard frames which had legacy
checkerboards with new frames. Also the multicamera transforms are
updated based on the new frames.