Unlock GIL in all releasable class methods
In d64366eb, the wrapper macros in Releasable.h were updated to unlock
the GIL when calling into the SDK. However, it seems a couple of macros
were missed in this commit. This meant that functions like
`Frame::point_cloud` and `Camera::state` could not run in parallel.

This commit adds the GIL unlocking to the wrapper macros that were
missed and also add the GIL unlock to one function that does not use
wrapper macros.