Version 2.17.0
In this version the C++ pybind11 wrapping code has been reorganized. The
`src` directory now contains two subdirectories `Core` and
`Visualization`. `Core` is all the wrapping code for the `ZividCore`
library (which previously was residing directly in the `src`
directory), and `Visualizer` contains the wrapping code for the
`ZividVisualization` library, which implements the `Visualizer` class
which is new in zivid-python in this release.

See the SDK changelog for details on what's new:
https://downloads.zivid.com/sdk/releases/2.17.0+5fc9f05e-1/Changelog.md